Initial setup (issues with header background-image)
For the header, I followed along with Tina and I found that the header background-image on it's own appears when added in CSS, however it then disappears completely when background-size & bakcground-postion are added in CSS (both when added individually and/or together). I'm not sure why this is as I've been following Tina step-by-step:

﻿ header {
    background-image: url('/images/the-sleep-co-4.jpg')
    background-size: cover;
    background-position: center;
 }﻿

﻿

However, when I create a class for the header bg-image and then apply it to the <header> in html, then the background image appears with the background-size and background-position settings working perfectly.